executor: split executor test into other package
