+++ pkg/executor/test/executor/BUILD.bazel
@@ -8,7 +8,7 @@ go_test(
         "main_test.go",
     ],
     flaky = True,
-    shard_count = 50,
+    shard_count = 48,
     deps = [
         "//pkg/config",
         "//pkg/config/kerneltype",
@@ -44,15 +44,12 @@ go_test(
         "//pkg/table/tables",
         "//pkg/tablecodec",
         "//pkg/testkit",
-        "//pkg/testkit/testdata",
         "//pkg/testkit/testfailpoint",
         "//pkg/types",
         "//pkg/util",
         "//pkg/util/dbterror/exeerrors",
         "//pkg/util/dbterror/plannererrors",
-        "//pkg/util/memory",
         "//pkg/util/mock",
-        "//pkg/util/replayer",
         "//pkg/util/rowcodec",
         "//pkg/util/sqlexec",
         "//pkg/util/timeutil",
