+++ pkg/executor/test/executor/executor_test.go
@@ -15,13 +15,10 @@
 package executor
 
 import (
-	"archive/zip"
 	"context"
 	"fmt"
 	"math"
 	"math/rand"
-	"os"
-	"path/filepath"
 	"reflect"
 	"runtime"
 	"strconv"
@@ -67,15 +64,12 @@ import (
 	"github.com/pingcap/tidb/pkg/table/tables"
 	"github.com/pingcap/tidb/pkg/tablecodec"
 	"github.com/pingcap/tidb/pkg/testkit"
-	"github.com/pingcap/tidb/pkg/testkit/testdata"
 	"github.com/pingcap/tidb/pkg/testkit/testfailpoint"
 	"github.com/pingcap/tidb/pkg/types"
 	"github.com/pingcap/tidb/pkg/util"
 	"github.com/pingcap/tidb/pkg/util/dbterror/exeerrors"
 	"github.com/pingcap/tidb/pkg/util/dbterror/plannererrors"
-	"github.com/pingcap/tidb/pkg/util/memory"
 	"github.com/pingcap/tidb/pkg/util/mock"
-	"github.com/pingcap/tidb/pkg/util/replayer"
 	"github.com/pingcap/tidb/pkg/util/rowcodec"
 	"github.com/pingcap/tidb/pkg/util/sqlexec"
 	"github.com/pingcap/tidb/pkg/util/timeutil"
@@ -84,151 +78,6 @@ import (
 	"github.com/tikv/client-go/v2/oracle"
 	"github.com/tikv/client-go/v2/testutils"
 )
-
-func checkFileName(s string) bool {
-	files := []string{
-		"config.toml",
-		"debug_trace/debug_trace0.json",
-		"meta.txt",
-		"stats/test.t_dump_single.json",
-		"schema/test.t_dump_single.schema.txt",
-		"schema/schema_meta.txt",
-		"table_tiflash_replica.txt",
-		"variables.toml",
-		"session_bindings.sql",
-		"global_bindings.sql",
-		"sql/sql0.sql",
-		"explain.txt",
-		"statsMem/test.t_dump_single.txt",
-		"sql_meta.toml",
-	}
-	for _, f := range files {
-		if strings.Compare(f, s) == 0 {
-			return true
-		}
-	}
-	return false
-}
-
-func TestPlanReplayer(t *testing.T) {
-	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/infoschema/mockTiFlashStoreCount", `return(true)`))
-	defer func() {
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/infoschema/mockTiFlashStoreCount"))
-	}()
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, b int, index idx_a(a))")
-	tk.MustExec("alter table t set tiflash replica 1")
-	tk.MustQuery("plan replayer dump explain select * from t where a=10")
-	defer os.RemoveAll(replayer.GetPlanReplayerDirName())
-	tk.MustQuery("plan replayer dump explain select /*+ read_from_storage(tiflash[t]) */ * from t")
-
-	tk.MustExec("create table t1 (a int)")
-	tk.MustExec("create table t2 (a int)")
-	tk.MustExec("create definer=`root`@`127.0.0.1` view v1 as select * from t1")
-	tk.MustExec("create definer=`root`@`127.0.0.1` view v2 as select * from v1")
-	tk.MustQuery("plan replayer dump explain with tmp as (select a from t1 group by t1.a) select * from tmp, t2 where t2.a=tmp.a;")
-	tk.MustQuery("plan replayer dump explain select * from t1 where t1.a > (with cte1 as (select 1) select count(1) from cte1);")
-	tk.MustQuery("plan replayer dump explain select * from v1")
-	tk.MustQuery("plan replayer dump explain select * from v2")
-	require.True(t, len(tk.Session().GetSessionVars().LastPlanReplayerToken) > 0)
-
-	// clear the status table and assert
-	tk.MustExec("delete from mysql.plan_replayer_status")
-	tk.MustQuery("plan replayer dump explain select * from v2")
-	token := tk.Session().GetSessionVars().LastPlanReplayerToken
-	rows := tk.MustQuery(fmt.Sprintf("select * from mysql.plan_replayer_status where token = '%v'", token)).Rows()
-	require.Len(t, rows, 1)
-}
-
-func TestPlanReplayerCaptureSEM(t *testing.T) {
-	originSEM := config.GetGlobalConfig().Security.EnableSEM
-	defer func() {
-		config.GetGlobalConfig().Security.EnableSEM = originSEM
-	}()
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("plan replayer capture '123' '123';")
-	tk.MustExec("create table t(id int)")
-	tk.MustQuery("plan replayer dump explain select * from t")
-	defer os.RemoveAll(replayer.GetPlanReplayerDirName())
-	tk.MustQuery("select count(*) from mysql.plan_replayer_status").Check(testkit.Rows("1"))
-}
-
-func TestPlanReplayerCapture(t *testing.T) {
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("plan replayer capture '123' '123';")
-	tk.MustQuery("select sql_digest, plan_digest from mysql.plan_replayer_task;").Check(testkit.Rows("123 123"))
-	tk.MustGetErrMsg("plan replayer capture '123' '123';", "plan replayer capture task already exists")
-	tk.MustExec("plan replayer capture remove '123' '123'")
-	tk.MustQuery("select count(*) from mysql.plan_replayer_task;").Check(testkit.Rows("0"))
-	tk.MustExec("create table t(id int)")
-	tk.MustExec("prepare stmt from 'update t set id = ?  where id = ? + 1';")
-	tk.MustExec("SET @number = 5;")
-	tk.MustExec("execute stmt using @number,@number")
-	_, sqlDigest := tk.Session().GetSessionVars().StmtCtx.SQLDigest()
-	_, planDigest := tk.Session().GetSessionVars().StmtCtx.GetPlanDigest()
-	tk.MustExec("SET @@tidb_enable_plan_replayer_capture = ON;")
-	tk.MustExec("SET @@global.tidb_enable_historical_stats_for_capture='ON'")
-	tk.MustExec(fmt.Sprintf("plan replayer capture '%v' '%v'", sqlDigest.String(), planDigest.String()))
-	err := dom.GetPlanReplayerHandle().CollectPlanReplayerTask()
-	require.NoError(t, err)
-	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/domain/shouldDumpStats", "return(true)"))
-	defer require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/domain/shouldDumpStats"))
-	tk.MustExec("execute stmt using @number,@number")
-	task := dom.GetPlanReplayerHandle().DrainTask()
-	require.NotNil(t, task)
-}
-
-func TestPlanReplayerContinuesCapture(t *testing.T) {
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-	tk := testkit.NewTestKit(t, store)
-
-	tk.MustExec("set @@global.tidb_enable_historical_stats='OFF'")
-	_, err := tk.Exec("set @@global.tidb_enable_plan_replayer_continuous_capture='ON'")
-	require.Error(t, err)
-	require.Equal(t, err.Error(), "tidb_enable_historical_stats should be enabled before enabling tidb_enable_plan_replayer_continuous_capture")
-
-	tk.MustExec("set @@global.tidb_enable_historical_stats='ON'")
-	tk.MustExec("set @@global.tidb_enable_plan_replayer_continuous_capture='ON'")
-
-	prHandle := dom.GetPlanReplayerHandle()
-	tk.MustExec("delete from mysql.plan_replayer_status;")
-	tk.MustExec("use test")
-	tk.MustExec("create table t(id int);")
-	tk.MustExec("set @@tidb_enable_plan_replayer_continuous_capture = 'ON'")
-	tk.MustQuery("select * from t;")
-	task := prHandle.DrainTask()
-	require.NotNil(t, task)
-	worker := prHandle.GetWorker()
-	success := worker.HandleTask(task)
-	defer os.RemoveAll(replayer.GetPlanReplayerDirName())
-	require.True(t, success)
-	tk.MustQuery("select count(*) from mysql.plan_replayer_status").Check(testkit.Rows("1"))
-}
-
-func TestPlanReplayerDumpSingle(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t_dump_single")
-	tk.MustExec("create table t_dump_single(a int)")
-	res := tk.MustQuery("plan replayer dump explain select * from t_dump_single")
-	defer os.RemoveAll(replayer.GetPlanReplayerDirName())
-	path := testdata.ConvertRowsToStrings(res.Rows())
-
-	reader, err := zip.OpenReader(filepath.Join(replayer.GetPlanReplayerDirName(), path[0]))
-	require.NoError(t, err)
-	defer func() { require.NoError(t, reader.Close()) }()
-	for _, file := range reader.File {
-		require.True(t, checkFileName(file.Name), file.Name)
-	}
-}
 
 func TestTimezonePushDown(t *testing.T) {
 	store := testkit.CreateMockStore(t)
@@ -753,32 +602,6 @@ func TestIssue19148(t *testing.T) {
 	require.Zero(t, tblInfo.Meta().Columns[0].GetFlag())
 }
 
-func TestOOMActionPriority(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t0")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("drop table if exists t2")
-	tk.MustExec("drop table if exists t3")
-	tk.MustExec("drop table if exists t4")
-	tk.MustExec("create table t0(a int)")
-	tk.MustExec("insert into t0 values(1)")
-	tk.MustExec("create table t1(a int)")
-	tk.MustExec("insert into t1 values(1)")
-	tk.MustExec("create table t2(a int)")
-	tk.MustExec("insert into t2 values(1)")
-	tk.MustExec("create table t3(a int)")
-	tk.MustExec("insert into t3 values(1)")
-	tk.MustExec("create table t4(a int)")
-	tk.MustExec("insert into t4 values(1)")
-	tk.MustQuery("select * from t0 join t1 join t2 join t3 join t4 order by t0.a").Check(testkit.Rows("1 1 1 1 1"))
-	action := tk.Session().GetSessionVars().StmtCtx.MemTracker.GetFallbackForTest(true)
-	// All actions are finished and removed.
-	require.Equal(t, action.GetPriority(), int64(memory.DefLogPriority))
-}
-
 // Test invoke Close without invoking Open before for each operators.
 func TestUnreasonablyClose(t *testing.T) {
 	store := testkit.CreateMockStore(t)
@@ -961,422 +784,6 @@ func TestTwiceCloseUnionExec(t *testing.T) {
 		exec.Close(e)
 		// No leak.
 	}
-}
-
-func TestPointGetPreparedPlan(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("drop database if exists ps_text")
-	defer tk.MustExec("drop database if exists ps_text")
-	tk.MustExec("create database ps_text")
-	tk.MustExec("use ps_text")
-
-	tk.MustExec(`create table t (a int, b int, c int,
-			primary key k_a(a),
-			unique key k_b(b))`)
-	tk.MustExec("insert into t values (1, 1, 1)")
-	tk.MustExec("insert into t values (2, 2, 2)")
-	tk.MustExec("insert into t values (3, 3, 3)")
-
-	pspk1Id, _, _, err := tk.Session().PrepareStmt("select * from t where a = ?")
-	require.NoError(t, err)
-	tk.Session().GetSessionVars().PreparedStmts[pspk1Id].(*plannercore.PlanCacheStmt).StmtCacheable = false
-	pspk2Id, _, _, err := tk.Session().PrepareStmt("select * from t where ? = a ")
-	require.NoError(t, err)
-	tk.Session().GetSessionVars().PreparedStmts[pspk2Id].(*plannercore.PlanCacheStmt).StmtCacheable = false
-
-	ctx := context.Background()
-	// first time plan generated
-	rs, err := tk.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(0))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(nil)
-
-	// using the generated plan but with different params
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(1))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 1 1"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(2))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("2 2 2"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk2Id, expression.Args2Expressions4Test(3))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("3 3 3"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk2Id, expression.Args2Expressions4Test(0))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(nil)
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk2Id, expression.Args2Expressions4Test(1))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 1 1"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk2Id, expression.Args2Expressions4Test(2))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("2 2 2"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk2Id, expression.Args2Expressions4Test(3))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("3 3 3"))
-
-	// unique index
-	psuk1Id, _, _, err := tk.Session().PrepareStmt("select * from t where b = ? ")
-	require.NoError(t, err)
-	tk.Session().GetSessionVars().PreparedStmts[psuk1Id].(*plannercore.PlanCacheStmt).StmtCacheable = false
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(1))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 1 1"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(2))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("2 2 2"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(3))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("3 3 3"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(0))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(nil)
-
-	// test schema changed, cached plan should be invalidated
-	tk.MustExec("alter table t add column col4 int default 10 after c")
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(0))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(nil)
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(1))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 1 1 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(2))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("2 2 2 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk2Id, expression.Args2Expressions4Test(3))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("3 3 3 10"))
-
-	tk.MustExec("alter table t drop index k_b")
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(1))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 1 1 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(2))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("2 2 2 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(3))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("3 3 3 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(0))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(nil)
-
-	tk.MustExec(`insert into t values(4, 3, 3, 11)`)
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(1))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 1 1 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(2))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("2 2 2 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(3))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("3 3 3 10", "4 3 3 11"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(0))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(nil)
-
-	tk.MustExec("delete from t where a = 4")
-	tk.MustExec("alter table t add index k_b(b)")
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(1))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 1 1 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(2))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("2 2 2 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(3))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("3 3 3 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, psuk1Id, expression.Args2Expressions4Test(0))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(nil)
-
-	// use pk again
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk2Id, expression.Args2Expressions4Test(3))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("3 3 3 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(3))
-	require.NoError(t, err)
-	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("3 3 3 10"))
-}
-
-func TestPointGetPreparedPlanWithCommitMode(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	setTxnTk := testkit.NewTestKit(t, store)
-	setTxnTk.MustExec("set global tidb_txn_mode=''")
-	tk1 := testkit.NewTestKit(t, store)
-	tk1.MustExec("drop database if exists ps_text")
-	defer tk1.MustExec("drop database if exists ps_text")
-	tk1.MustExec("create database ps_text")
-	tk1.MustExec("use ps_text")
-
-	tk1.MustExec(`create table t (a int, b int, c int,
-			primary key k_a(a),
-			unique key k_b(b))`)
-	tk1.MustExec("insert into t values (1, 1, 1)")
-	tk1.MustExec("insert into t values (2, 2, 2)")
-	tk1.MustExec("insert into t values (3, 3, 3)")
-
-	pspk1Id, _, _, err := tk1.Session().PrepareStmt("select * from t where a = ?")
-	require.NoError(t, err)
-	tk1.Session().GetSessionVars().PreparedStmts[pspk1Id].(*plannercore.PlanCacheStmt).StmtCacheable = false
-
-	ctx := context.Background()
-	// first time plan generated
-	rs, err := tk1.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(0))
-	require.NoError(t, err)
-	tk1.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(nil)
-
-	// using the generated plan but with different params
-	rs, err = tk1.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(1))
-	require.NoError(t, err)
-	tk1.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 1 1"))
-
-	// next start a non autocommit txn
-	tk1.MustExec("set autocommit = 0")
-	tk1.MustExec("begin")
-	// try to exec using point get plan(this plan should not go short path)
-	rs, err = tk1.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(1))
-	require.NoError(t, err)
-	tk1.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 1 1"))
-
-	// update rows
-	tk2 := testkit.NewTestKit(t, store)
-	tk2.MustExec("use ps_text")
-	tk2.MustExec("update t set c = c + 10 where c = 1")
-
-	// try to point get again
-	rs, err = tk1.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(1))
-	require.NoError(t, err)
-	tk1.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 1 1"))
-
-	// try to update in session 1
-	tk1.MustExec("update t set c = c + 10 where c = 1")
-	err = tk1.ExecToErr("commit")
-	require.True(t, kv.ErrWriteConflict.Equal(err), fmt.Sprintf("error: %s", err))
-
-	// verify
-	rs, err = tk1.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(1))
-	require.NoError(t, err)
-	tk1.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 1 11"))
-
-	rs, err = tk1.Session().ExecutePreparedStmt(ctx, pspk1Id, expression.Args2Expressions4Test(2))
-	require.NoError(t, err)
-	tk1.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("2 2 2"))
-
-	tk2.MustQuery("select * from t where a = 1").Check(testkit.Rows("1 1 11"))
-}
-
-func TestPointUpdatePreparedPlan(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("drop database if exists pu_test")
-	defer tk.MustExec("drop database if exists pu_test")
-	tk.MustExec("create database pu_test")
-	tk.MustExec("use pu_test")
-
-	tk.MustExec(`create table t (a int, b int, c int,
-			primary key k_a(a),
-			unique key k_b(b))`)
-	tk.MustExec("insert into t values (1, 1, 1)")
-	tk.MustExec("insert into t values (2, 2, 2)")
-	tk.MustExec("insert into t values (3, 3, 3)")
-
-	updateID1, pc, _, err := tk.Session().PrepareStmt(`update t set c = c + 1 where a = ?`)
-	require.NoError(t, err)
-	tk.Session().GetSessionVars().PreparedStmts[updateID1].(*plannercore.PlanCacheStmt).StmtCacheable = false
-	require.Equal(t, 1, pc)
-	updateID2, pc, _, err := tk.Session().PrepareStmt(`update t set c = c + 2 where ? = a`)
-	require.NoError(t, err)
-	tk.Session().GetSessionVars().PreparedStmts[updateID2].(*plannercore.PlanCacheStmt).StmtCacheable = false
-	require.Equal(t, 1, pc)
-
-	ctx := context.Background()
-	// first time plan generated
-	rs, err := tk.Session().ExecutePreparedStmt(ctx, updateID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 4"))
-
-	// using the generated plan but with different params
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updateID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 5"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updateID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 6"))
-
-	// updateID2
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updateID2, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 8"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updateID2, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 10"))
-
-	// unique index
-	updUkID1, _, _, err := tk.Session().PrepareStmt(`update t set c = c + 10 where b = ?`)
-	require.NoError(t, err)
-	tk.Session().GetSessionVars().PreparedStmts[updUkID1].(*plannercore.PlanCacheStmt).StmtCacheable = false
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updUkID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 20"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updUkID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 30"))
-
-	// test schema changed, cached plan should be invalidated
-	tk.MustExec("alter table t add column col4 int default 10 after c")
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updateID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 31 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updateID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 32 10"))
-
-	tk.MustExec("alter table t drop index k_b")
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updUkID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 42 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updUkID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 52 10"))
-
-	tk.MustExec("alter table t add unique index k_b(b)")
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updUkID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 62 10"))
-
-	rs, err = tk.Session().ExecutePreparedStmt(ctx, updUkID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 72 10"))
-
-	tk.MustQuery("select * from t where a = 1").Check(testkit.Rows("1 1 1 10"))
-	tk.MustQuery("select * from t where a = 2").Check(testkit.Rows("2 2 2 10"))
-}
-
-func TestPointUpdatePreparedPlanWithCommitMode(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-
-	setTxnTk := testkit.NewTestKit(t, store)
-	setTxnTk.MustExec("set global tidb_txn_mode=''")
-	tk1 := testkit.NewTestKit(t, store)
-	tk1.MustExec("drop database if exists pu_test2")
-	defer tk1.MustExec("drop database if exists pu_test2")
-	tk1.MustExec("create database pu_test2")
-	tk1.MustExec("use pu_test2")
-
-	tk1.MustExec(`create table t (a int, b int, c int,
-			primary key k_a(a),
-			unique key k_b(b))`)
-	tk1.MustExec("insert into t values (1, 1, 1)")
-	tk1.MustExec("insert into t values (2, 2, 2)")
-	tk1.MustExec("insert into t values (3, 3, 3)")
-
-	ctx := context.Background()
-	updateID1, _, _, err := tk1.Session().PrepareStmt(`update t set c = c + 1 where a = ?`)
-	tk1.Session().GetSessionVars().PreparedStmts[updateID1].(*plannercore.PlanCacheStmt).StmtCacheable = false
-	require.NoError(t, err)
-
-	// first time plan generated
-	rs, err := tk1.Session().ExecutePreparedStmt(ctx, updateID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk1.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 4"))
-
-	rs, err = tk1.Session().ExecutePreparedStmt(ctx, updateID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk1.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 5"))
-
-	// next start a non autocommit txn
-	tk1.MustExec("set autocommit = 0")
-	tk1.MustExec("begin")
-	// try to exec using point get plan(this plan should not go short path)
-	rs, err = tk1.Session().ExecutePreparedStmt(ctx, updateID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk1.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 6"))
-
-	// update rows
-	tk2 := testkit.NewTestKit(t, store)
-	tk2.MustExec("use pu_test2")
-	tk2.MustExec(`prepare pu2 from "update t set c = c + 2 where ? = a "`)
-	tk2.MustExec("set @p3 = 3")
-	tk2.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 5"))
-	tk2.MustExec("execute pu2 using @p3")
-	tk2.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 7"))
-	tk2.MustExec("execute pu2 using @p3")
-	tk2.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 9"))
-
-	// try to update in session 1
-	tk1.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 6"))
-	err = tk1.ExecToErr("commit")
-	require.True(t, kv.ErrWriteConflict.Equal(err), fmt.Sprintf("error: %s", err))
-
-	// verify
-	tk2.MustQuery("select * from t where a = 1").Check(testkit.Rows("1 1 1"))
-	tk1.MustQuery("select * from t where a = 2").Check(testkit.Rows("2 2 2"))
-	tk2.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 9"))
-	tk1.MustQuery("select * from t where a = 2").Check(testkit.Rows("2 2 2"))
-	tk1.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 9"))
-
-	// again next start a non autocommit txn
-	tk1.MustExec("set autocommit = 0")
-	tk1.MustExec("begin")
-	rs, err = tk1.Session().ExecutePreparedStmt(ctx, updateID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk1.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 10"))
-
-	rs, err = tk1.Session().ExecutePreparedStmt(ctx, updateID1, expression.Args2Expressions4Test(3))
-	require.Nil(t, rs)
-	require.NoError(t, err)
-	tk1.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 11"))
-	tk1.MustExec("commit")
-
-	tk2.MustQuery("select * from t where a = 3").Check(testkit.Rows("3 3 11"))
 }
 
 func TestApplyCache(t *testing.T) {
